Extended Travel time analysis (#1763)
* new analysis class to download travel time patterns

* Renaming to avoid some confusion (#1736)

The short trips are now renamed to normal trip (i.e. the trips within the city or region, which is the main type of the PT trips in the simulation).

The long trips are now renamed to long distance trip (i.e. intercity trips). Usually the ticket for those trips are different from the normal trips. Therefore a different fare parameter is used for those trips.

* Add VRPTransportCostsFactory

* Travel time validation: add short delay between 2 URL requests (#1739)

Add a short delay between 2 URL request to avoid sending too many request to the online API at the same time (Error code 429)

* fix typo

* add a new route validator

* add some parsing code to create a network

* write a standard network with free flow speeds

* improve speed recognition

* implemented simple network route validator

* write network change events for time-dependent network

* new analysis class to download travel time patterns

* fix typo

* add a new route validator

* add some parsing code to create a network

* write a standard network with free flow speeds

* improve speed recognition

* implemented simple network route validator

* write network change events for time-dependent network

* implemented use of network change events

* add consttructors to the csv options

* add test for csv options

* enforce commons-codec version

Co-authored-by: Chengqi Lu <43133404+luchengqi7@users.noreply.github.com>
Co-authored-by: steffenaxer <26229392+steffenaxer@users.noreply.github.com>