Update Array based travel time data
The travel time observer does not return the correct value for the travel time for the unused links. This difference is due to the discrete QSim time step. This fix incorporate the QSim config in the Travel Time data construction, such that the travel time for the unused link can be calculated more accurately.