Update VehicleDataEntryFactoryImpl.java (#2106)
With this update, the schedule of a DVRP can end with different types of task.

This is necessary when we want to have partial DVRP schedule during the day. By not computing the full schedule whenever the schedule gets update, we can save some computational workload. This can also be helpful for customize DRT optimizer.