I tried to comment on the changes that were made so far.
I am almost certain that adding the grade into Hbefa(Warm)EmissionFactor is the wrong design decision, since so far this class contains hbefa output, and grade is input for the lookup.  I thus removed it there.  It was also not made consistent with the test cases, so rolling it back is better.

There are no tests for using the grades.  There was something programmed in a fork of matsim-code-examples, but it was just trying out if it ran, but no results.  I am also somewhat confident that that attempt at a test case did not contain corresponding non-zero grade entries in hbefa input files.

In my intuition, the current version is (considerably) less than 50% towards having grades in emissions, i.e. still a long way to go.