core: remove flaky MemoryObserverTest
It has recently failed a couple of times in GitHub Actions (linux):
```
Error: MemoryObserverTest.testStartStop:59 [Number of log messages]
Expecting actual: 6 to be between: [2, 4]
```
In the past it also failed on AppVeyor (windows) and had to be adapted.

It seems that any non-negative number is theoretically possible (in all assertions).