Update max detour constraint
We now make use the initial planned pick up time (+ a small buffer) + max ride duration to determine the latest arrival time. This will also be a fixed value, therefore, it will not impact the matching logic. The latest arrival time based on the original alpha beta value is also included here (with Math.min operation).

The main change here is that we can no longer insert a lot of stops before the scheduled pick ups, as we limit the latest pick up time based on the initial scheduled pick up time. This should also be meaningful in the reality.