Update MATSim CW26 (#104)
* CreateDrtDashboard

* rename pooling ratio to occupancy rate

* account for group bookings in DrtAnalysisPostProcessing (and thus DrtDashboard)

* typo in CreateDrtDashboard

* fix wrong sum calculation

* debug NoiseAnalysis and proper testing in NoiseDashboardTests

* fix wrong map selection

* add reduction of numberOfJobs

* add comments and format log massages

* combine

* rename parameter

* reduce code

* format logging

* use method for demand calculating

* use method for shipment creation

* remove todo

* use method to calculate demand based on link lengths

* introduce parameter

* simplify

* cleanup

* Support TransitRouteStops where boarding or alighting is not allowed

In reality, there are trains that call at stops but where it might not be allowed to enter or exit such trains. Examples are night trains, where on the first few stops, only boarding is allowed, while at the last few stops only alighting is allowed. This means that it is not allowed to use such night trains to just travel between the first few stops.

This commit adds two optional attributes `allowBoarding` and `allowAlighting` to TransitRouteStops, with the default value being true for backwards compatibility and when the attribute is not specified. SwissRailRaptor respects these attributes and will not return routes where agents would enter a vehicle at a route stop where boarding is not allowed, or where agents would exit a vehicle at a route stop where alighting is not allowed.

* make drt shifts multi mode drt comatible

* adjust shift efficiency tracker as well

---------

Co-authored-by: schlenther <schlenther@vsp.tu-berlin.de>
Co-authored-by: Ricardo Ewert <ewert@vsp.tu-berlin.de>
Co-authored-by: Ricardo Ewert <60344745+rewertvsp@users.noreply.github.com>
Co-authored-by: Marcel Rieser <rieser@simunto.com>