Fix storage capacity on SplitOutLinks:
When using 'kinematicWaves', the storage capacity of the link is increased to have enough room for backwards travelling holes. This increase was not reflected by SplitOutLinks which mirror the storage capacity of the downstream SplitInLink. Changed initialization of storage capacities, so that they are assigned max capacities directly. The calculations to arrive at the capacity are moved into static methods.