Store the pcuScalingFactor as factor instead of pre-computing the scaled pcu
Apparently, the FISS module changes the pcu of vehicles dynamically inbetween iterations. At the same time, vehicles are only generated once before the simulation starts. When pre-computing the scaled pcu, this change is not reflected by the vehicle. Now, the pcu is scaled in the getter directly, based on the pcu in the vehicle type. This way, dynamic changes to the vehicle type's pcu are refelcted during the simulation