contrib(profiling): WIP JFRStopwatch - statistics about method sampling
issues:
- negative values?
  - iterationEndsListeners sometimes start later than the recorded iteration end?
  - and may also cause stack exceptions in the stopwatch

Todo:
Try to check why the times from the JFRIterationEvents differ that much from the execution samples.
If inexplicable, maybe ignore the iteration event times and detect iteration start and end
by checking if all operations were recorded or use the begin of the next iteration as end for the previous one.