contrib(profiling): really fix iteration end listener priorities
While Double.MAX_VALUE is the greatest absolute value for a double, Double.MIN_VALUE is not the negative equivalent.
It is the smallest positive value.

Additionally, due to ieee precision inaccuracies, you cannot simply add a positive integer and assume that it differs from the previous value.
-> Double.MAX_VALUE == Double.MAX_VALUE - 100

You would need to calculate how much exactly you need to subtract to reach the next smallest number that can be represented in Double precision.
While Integer would be the better representation of a priority by not suffering from precision errors,
we use MAX_VALUE and INFINITY to clearly and easily differentiate between two values.