contrib(profiling): refactor the abstract profiling aspect into an @Aspect
Apparently, it is not possible to mix the .aj aspects and @Aspects via inheritance.
Using the annotations is a bit more verbose, but also much more similar to the existing code base.
While we can keep other examples of defining aspects as .aj files, this one is intended to be used more,
so we switch it to the @Aspect notation.